Make SSL optional when cert is set for ingestor and kibana
This introduces `properties.haproxy.kibana.enable_ssl` and
`properties.haproxy.ingestor.enable_ssl` (both defaults to false)
which take care of enable SSL on the haproxy more flexible then
before.

Previously if a SSL cert was given, port `5514` (ingestor) and
kibana, port 80 (HTTP) or 443 (HTTPS) was automatically enabled.

With this commit, one can be enabled when the other is disabled.

Example:
Ingestor (syslog) traffic should be plainext, but the
Kibana traffic should be encrypted.