Add properties.haproxy.elastic.enable_ssl to specfile
The properties.haproxy.elastic.enable_ssl property was missing in
the spec file as @Infra-Red pointed out correctly.

Also the `default` values for `enable_ssl` have been changed to
true to be backward compatible. If an ssl cert is set, ssl will be
enabled on all the services on ha_proxy but can be disabled by
`enable_ssl: false`.